requirements.txt
python-dateutil
tabulate
pdf2image
pytesseract
